+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change log
 
 For detailed information check [Blogophon's releases on Github](https://github.com/fboes/blogophon/releases).
 
+* [x] Image sublines for images
+
 1.2.0
 -----